Refactor project structure and update imports
Moved core modules and models into a new 'src' package for better organization. Updated all import statements to use the new package structure. Added a project structure section to the README for improved documentation and clarity.
