+++ src/string_util.py
